Added stats to Player and Mob, Replaced tree with pillar
Added the 5 stats to the player and mob, with the mob also getting the type stat. Additionally In order to make it more of a simple and less corporate style the tree png was replaced with a black color rect object with a white polygon2d object in front of it, in order to have a white box with a black outline.