guns now spawn at specific marker2ds
Now when a gun is added it checks how many techniques currently exist and adds the gun to the correct marker2d. Will need to redo UI in future so that guns do not cover UI and vice versa. Will soon merge this branch into main and start working on a more complete level up system.