added in healing bullet and various modifiers
Healing bullet shoots out a bullet and if it's the same type as the caster it collides with it, and heals the target for a bit of hp. It's honestly a bit busted.

Also added the quick modifier, which decreases the cooldown at the cost of increased cost

the cheap modifier which decreases cost in exchange for lesser value

And the self modifier which increases the value in exchange for it only pointing at your self.